Add files via upload
This is a script that allows the robot to scan barcodes, and when the output excel is then ejected, The scripts loops over the columns and asks for user input for any missed barcodes. It was crucial to maintain the order of the barcodes as this was used for a medical samples for people that was correlated in a different program based on its' location in the plate.